Add a functional system shuffler generator (#3)
Works almost the same as the original System Shuffler plugin, but with some fixes and niceties.
The Rust code could be less in quantity, but my focus was on getting it working.
The generated plugin does its best to minimize errors in `errors.txt`.  There's only so much I can do about it.