Support `link` and `unlink` nodes in events, and trigger side events immediately
This is kind of a nightmare commit because in my search for satisfaction I lost control.

The driving realization for this was remembering I forgot `link` and `unlink` nodes in events.
Those are somewhat common for changing the universe, so ideally they'd be found and have shuffled versions.

While I was at it, I tried to make the side events in each `universe_preset_N/events.txt` trigger immediately.
This way the player won't be required to land for an up-to-date layout unless the change happens after the shuffle.

Side events aren't working as they should.
I'm testing with a save that started without the plugin but already has the Pug invasion.
This should be fixed before merging.