Support `link` and `unlink` nodes in events, and trigger side events … (#4)
* Support `link` and `unlink` nodes in events, and trigger side events immediately

This is kind of a nightmare commit because in my search for satisfaction I lost control.

The driving realization for this was remembering I forgot `link` and `unlink` nodes in events.
Those are somewhat common for changing the universe, so ideally they'd be found and have shuffled versions.

While I was at it, I made the side events in each `universe_preset_N/events.txt` trigger immediately.
This way the player won't be required to land for an up-to-date layout unless the change happens after the shuffle.

* Miscellaneous fixes

- rename last shuffle condition
- use the proper condition for last shuffle days in mission `System Shuffler: Select Preset`
- call `conditional_events` in job `System Shuffler: Restore Universe` now that side events attempt to run immediately
- prefix side event missions with "zzzzz" in hopes of making them trigger last
- side event missions now check `CURRENT_PRESET` in their offer condition
- remove custom jump ranges with `"jump range" 0` instead of `remove "jump range"`
- `INSTALLED` gets set from jobs, too

* Fix logic errors (it works now)

Also prefixed missions with "zzzzz" and made all offer precedences -1000000